Allow unit tests to bypass mocks if apiPrincipal is set
With the changes in this commit, the unit tests will allow TheHive
service to make real requests to TheHive server if the apiPrincipal is
pre-configured in the dxlthehiveservice.config file. If the apiPrincipal
is not pre-configured, the tests will continue to use mocks. Assertions
against response data were also expanded to validate more of the data
that the samples use.