Modify sample test config file logic to support Windows
Previously, the sample tests flushed but did not close the temporary
config file that they create when tests are running against a mock
TheHive server. On Linux-based systems, the tests were able to read and
use the config file. On Windows systems, however, the tests would be
unable to open the config file for reading because it the file was
already open - leading the tests to fail when run on Windows.

This commit reworks the logic to ensure that the file has been closed
before the service tries to read it during startup. This change allows
the tests to pass both on Linux-based and Windows systems.